32bit ready

+++ headers/utils.h
@@ -3,14 +3,16 @@
 
 #include <string>
 #include <vector>
+#include <main_types.h>
+
 using namespace std;
 
 void split(const string &str, char c, vector<string> &elements);
-bool inputToNumber(string input, unsigned long long int& number);
+bool inputToNumber(string input, BigInt& number);
 string toUppercase(string input);
 bool endsWith(std::string const &value, std::string const &ending);
 string executeCommand(const char* cmd);
-unsigned long long int invertEndian(unsigned long long int val);
+BigInt invertEndian(BigInt val);
 string removeConsoleChars(string consoleOut);
 
 #endif
